tui: simplify startup test coverage
